fix(release): address MCP packaging review feedback
