+++ .github/workflows/release-prep.yml
@@ -110,6 +110,7 @@ jobs:
           git push --follow-tags origin HEAD:"$DEFAULT_BRANCH"
 
       - name: Create draft GitHub release
+        id: draft_release
         env:
           GH_TOKEN: ${{ secrets.RELEASE_AUTOMATION_TOKEN }}
           VERSION: ${{ inputs.version }}
@@ -122,10 +123,12 @@ jobs:
               echo "Release $TAG already exists and is not a draft. Refusing to overwrite a published release." >&2
               exit 1
             fi
+            echo "status=verified" >> "$GITHUB_OUTPUT"
             exit 0
           fi
 
           gh release create "$TAG" --draft --title "OpenDucktor v$VERSION" --generate-notes
+          echo "status=created" >> "$GITHUB_OUTPUT"
 
       - name: Dispatch desktop publish workflow
         env:
@@ -143,6 +146,7 @@ jobs:
 
       - name: Publish workflow summary
         env:
+          DRAFT_RELEASE_STATUS: ${{ steps.draft_release.outputs.status }}
           VERSION: ${{ inputs.version }}
         run: |
           {
@@ -150,6 +154,6 @@ jobs:
             echo
             echo "- Release commit created on default branch"
             echo "- Release tag pushed: v$VERSION"
-            echo "- Draft GitHub release created with generated notes"
+            echo "- Draft GitHub release ${DRAFT_RELEASE_STATUS:-verified} with generated notes"
             echo "- Dispatched workflows: Release Desktop, Publish MCP Package"
           } >> "$GITHUB_STEP_SUMMARY"
